Solve pod lint warning issue

+++ AppDevPods/AppDevImageKit/UIImage+ADKImageFilter.m
@@ -99,7 +99,7 @@
     effectOutBuffer.rowBytes = CGBitmapContextGetBytesPerRow(effectOutContext);
 
     CGFloat inputRadius = blurRadius * [[UIScreen mainScreen] scale];
-    NSUInteger radius = floor(inputRadius * 3. * sqrt(2 * M_PI) / 4 + 0.5);
+    uint32_t radius = floorl(inputRadius * 3. * sqrt(2 * M_PI) / 4 + 0.5);
     if (radius % 2 != 1) {
         radius += 1;
     }
